feat: Support multiple validation sets
This change allows for multiple validation datasets to be used during training.
The evaluation logic is updated to handle and report results for each validation
dataset individually, as well as an aggregated view. The `GPTDatasetConfig`
is updated with a `multiple_validation_sets` flag to enable this feature.
The data loader and iterator building functions are also modified to return
lists of dataloaders and iterators when multiple validation sets are configured.

Co-authored-by: hawkoli1987 <hawkoli1987@gmail.com>
Signed-off-by: yuli <hawkoli1987@gmail.com>